+++ packages/boustro_starter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **Changed**: Image embed overlay buttons no longer have a fade effect.
+
 ## [0.1.0]
 
 Initial publish.